DPR-564 moved all sorting logic to repository from service and refactored the tests
